[FE-Seller][Design] 프로필 UI 수정

See merge request s10-final/S10P31D104!254

+++ frontend-seller/src/service/menu/DeleteMenuAPI.ts
@@ -5,9 +5,14 @@ const DeleteMenuAPI = () => {
   const cookies = new Cookies()
   const accessToken = cookies.get("accessToken")
   const DeleteMenu = (menuId:any) => {
-    axios.delete(import.meta.env.VITE_BASE_URL + `/api/menus/${menuId}`, {
-      withCredentials: true,
-      params: {
+    const isConfirm = confirm(
+      "메뉴를 삭제하시겠습니까?"
+    );
+    if (isConfirm) {
+
+      axios.delete(import.meta.env.VITE_BASE_URL + `/api/menus/${menuId}`, {
+        withCredentials: true,
+        params: {
         "menu-id": menuId
       },
       headers: {
@@ -26,6 +31,7 @@ const DeleteMenuAPI = () => {
       console.log(err)
       window.alert("메뉴 삭제 실패")
     })
+  }
   };
   return {
     DeleteMenu
